Initial version of the Node Categories addon.
This adds a basic panel to the new node toolbar region, from which a number of panels can be created. Users can add node categories, which are stored (for the time being) in the scene. These categories are then in turn used to automatically generate a number of Panel subclasses. Node types can be added as 'items' in each category, which will result in a operator button to add a node of that type. In addition to the basic node type, a number of settings can be registered for each item to initialize the new node.

For editing the categories the user must first enable the 'Edit' button in the main panel, this is a sort of "edit mode" to prevent cluttered UI.

Addon is marked as TESTING for now, there are a few issues that need to be solved and it's not feature complete yet.



git-svn-id: https://svn.blender.org/svnroot/bf-extensions/contrib/py/scripts/addons@4042 96b784f8-4ae3-dd11-aff7-0030488c597c